make shared_modules_path optional

+++ utils/preparetemp.go
@@ -22,12 +22,15 @@ func (tofuguStruct *Tofugu) PrepareTemp() {
 		log.Fatalf("failed to rsync tofi to tempdir %s\n", err)
 	}
 
-	command = exec.Command("ln", "-sf", tofuguStruct.SharedModulesPath, cmdTempDirFullPath)
-	output, err = command.CombinedOutput()
-	if err != nil {
-		os.RemoveAll(cmdTempDirFullPath)
-		log.Printf("failed %s", output)
-		log.Fatalf("failed symlink shared_modules to tempdir %s\n", err)
+	if tofuguStruct.SharedModulesPath != "" {
+		command = exec.Command("ln", "-sf", tofuguStruct.SharedModulesPath, cmdTempDirFullPath)
+		output, err = command.CombinedOutput()
+		if err != nil {
+			os.RemoveAll(cmdTempDirFullPath)
+			log.Printf("failed %s", output)
+			log.Fatalf("failed symlink shared_modules to tempdir %s\n", err)
+		}
+		log.Println("TofuGu symlinked shared_modules to tempdir : " + tofuguStruct.SharedModulesPath)
 	}
 
 	tofuguStruct.CmdWorkTempDir = cmdTempDirFullPath
